Match BgClipCheckTouchingTransitionOrTank

+++ include/connection.h
@@ -9,7 +9,7 @@ void ConnectionUpdateHatches(void);
 void ConnectionUpdateHatchAnimation(u8 dontSetRaw, u32 hatch);
 void ConnectionHatchFlashingAnimation(u8 hatch);
 void ConnectionOverrideOpenedHatch(u8 hatch, u32 type);
-u8 ConnectionCheckEnterDoor(u16 yPosition, u16 xPosition);
+u32 ConnectionCheckEnterDoor(u16 yPosition, u16 xPosition);
 u8 ConnectionCheckAreaConnection(u16 yPosition, u16 xPosition);
 void ConnectionProcessDoorType(u8 type);
 u8 ConnectionFindEventBasedDoor(u8 sourceDoor);
